Mainly fixed 2 things on "On Display Messages": -They might have never drawn if DrawMessages wasn't called before they actually expired. -Their fade was wrong if the duration of the message was less than the fade time.
This makes them much more useful for debugging, I know there might be other means of debugging like logs and imgui, but this was the simplest so that's what I used.
If you want to print the same message every frame, but with a slightly different value to see the changes, it now work.

To compensate for the fact that they are now always rendered once, so on start up a lot of old messages (printed while the emulation was off) could show up, I've added a "drop" time, which means if a msg isn't rendered for the first time within that time, it will be dropped and never rendered.