InputCommon: add devices sorting order
This is to make sure the MouseAndKeyboard (Windows) device is always added first,
effectively making it the "default" default device (index 0), otherwise, with the recent
changes, which allowed it to be added async, it could end up in any place in the list.
Also make hotplug devices like DSU and Wiimote appear at the end of the list.