InputCommon: fix small bug which could prevent a DInput Joystick device from being added
If the device first attempted creation without a window handle, which means it couldn't obtain exclusive mode, which implies it would have not outputs, it would not be created as a device if it also had no valid inputs, but it would be added to the list of GUIDs in use.
If we then re-populated devices with a valid window handle, this device could never be actually recreated, even if it now had some outputs.