InputCommon: finalize ControllerInterface safety and stability refactor
-All old devices shared pts are now released immediately (from UI and ControlReference Expressions)
when we change the devces list. This fixes a bug where devices could not be recreated because
an old version of them still existed as a shared ptr. This also means we basically don't really
have any relevant reason to keep devices as shared ptrs, but it's also NOT a problem anymore.
-A lot of platforms now populate asyncronously, without stalling the main thread.
-Qt UI "Refresh Devices" button doesn't stop the CPU thread anymore (it's not necessary anymore)
-Fixes add/remove order of devices population, and rare cases where devices could be added twice or be missing
-Prevent devices from being added before the first population (RefreshDevices()),
it was literally pointless as they would have been cleaned anyway.
-Implemented a quick DevicesRefresh() which only recreates some devices if
only the render widget had changed (workaround for issue 11702 and 12478).
-Avoid broadcasting DevicesChanged() callbacks when there were no changes.
-Avoid refreshing devices on shutdown.
-Small fixes/polish, improve thread safety
-Removed window from OSX backend as it wasn't used
-Improved code to handle long waits for events that might not come (to fix the hangs)